+++ src/model/collision.cpp
@@ -29,7 +29,7 @@ bool Collision::load(Model *const model, String filePath)
 		return false;
 	}
 
-	const size_t fileSize = file->size();
+	const size_t fileSize = static_cast<size_t>(file->size());
 	UniquePtr<uint8_t[]> buffer(new uint8_t[fileSize]);
 	file->read((char *)buffer.get(), sizeof(char), fileSize);
 	file.reset();
