+++ src/structs/dds.cpp
@@ -28,7 +28,7 @@ namespace dds
 		}
 		else
 		{
-			const size_t fileSize = file->size();
+			const size_t fileSize = static_cast<size_t>(file->size());
 			UniquePtr<uint8_t[]> buffer(new uint8_t[fileSize]);
 			file->read((char *)buffer.get(), sizeof(char), fileSize);
 			file.reset();
